requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
motor==3.0.0
multidict==6.0.2
pymongo==4.1.1
python-dotenv==0.20.0
pytz==2022.1
watchdog==2.1.8
yarl==1.7.2
